Move 3 integration tests to specific go build tags and change test name Prefix
