+++ operator/internal/decommissioning/statefulset_decommissioner_test.go
@@ -7,12 +7,15 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+//go:build integration
+
 package decommissioning_test
 
 import (
 	"context"
 	_ "embed"
 	"encoding/json"
+	"github.com/redpanda-data/redpanda-operator/pkg/testutil"
 	"testing"
 	"time"
 
@@ -43,7 +46,11 @@ import (
 //go:embed role.yaml
 var decommissionerRBAC []byte
 
-func TestStatefulSetDecommissioner(t *testing.T) {
+func TestIntegrationStatefulSetDecommissioner(t *testing.T) {
+	t.Skip("Too flaky")
+	testutil.SkipIfNotIntegration(t)
+	testutil.RequireTimeout(t, time.Minute*10)
+
 	if testing.Short() {
 		t.Skip("skipping long running test as -short was specified")
 	}
